shared/requirements.txt
# Shared package dependencies
# This file contains all dependencies needed by the shared package

# Web framework
fastapi>=0.100.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# HTTP client
httpx>=0.24.0

# JWT handling
PyJWT>=2.8.0

# Structured logging
structlog>=23.1.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.28.0

# AWS SDK
boto3>=1.34.0

# Type checking (dev)
mypy>=1.5.0
types-PyJWT>=1.7.1
